refactor: split VM instance monolith into focused files, eliminate launchtemplate.Provider
Split the 1,089-line vminstance.go monolith into focused files matching the
AKS Machine path pattern:

- vminstance.go (~583 lines): VMProvider interface, struct, constructor,
  BeginCreate, Get, List, Delete, Update, NIC CRUD, promise, beginLaunchInstance
- vminstancehelpers.go (~402 lines): buildVMTemplate() + configure* helpers +
  resolveBootstrapAndImageData() + extension spec builders
- vminstanceutils.go (~198 lines): BuildNodeClaimFromVM(), utility functions,
  priority mappings, extension helpers
- vminstancenic.go (~173 lines): NIC construction with buildAndCreateNIC()
  orchestrator

Key changes:
- Eliminate launchtemplate.Provider intermediate abstraction; inline its logic
  into DefaultVMProvider via resolveBootstrapAndImageData()
- Replace 7-layer parameter threading (options → StaticParameters → Parameters
  → Template → createVMOptions → Azure API) with direct configure* helpers
- Move BuildNodeClaimFromVM from cloudprovider to instance package (parallels
  BuildNodeClaimFromAKSMachine)
- NIC code takes subnetID/tags directly instead of *launchtemplate.Template
- Delete pkg/providers/launchtemplate/launchtemplate.go (tags.go and
  parameters/ remain as shared code)
