+++ pkg/test/environment.go
@@ -42,7 +42,6 @@ import (
 	"github.com/Azure/karpenter-provider-azure/pkg/providers/instance"
 	"github.com/Azure/karpenter-provider-azure/pkg/providers/instancetype"
 	"github.com/Azure/karpenter-provider-azure/pkg/providers/kubernetesversion"
-	"github.com/Azure/karpenter-provider-azure/pkg/providers/launchtemplate"
 	"github.com/Azure/karpenter-provider-azure/pkg/providers/loadbalancer"
 	"github.com/Azure/karpenter-provider-azure/pkg/providers/networksecuritygroup"
 	"github.com/Azure/karpenter-provider-azure/pkg/providers/pricing"
@@ -99,7 +98,6 @@ type Environment struct {
 	KubernetesVersionProvider    kubernetesversion.KubernetesVersionProvider
 	ImageProvider                imagefamily.NodeImageProvider
 	ImageResolver                imagefamily.Resolver
-	LaunchTemplateProvider       *launchtemplate.Provider
 	LoadBalancerProvider         *loadbalancer.Provider
 	NetworkSecurityGroupProvider *networksecuritygroup.Provider
 
@@ -168,20 +166,6 @@ func NewRegionalEnvironment(ctx context.Context, env *coretest.Environment, regi
 		pricingProvider,
 		unavailableOfferingsCache)
 	imageFamilyResolver := imagefamily.NewDefaultResolver(env.Client, imageFamilyProvider, instanceTypesProvider, nodeBootstrappingAPI)
-	launchTemplateProvider := launchtemplate.NewProvider(
-		ctx,
-		imageFamilyResolver,
-		imageFamilyProvider,
-		lo.ToPtr("ca-bundle"),
-		testOptions.ClusterEndpoint,
-		"test-tenant",
-		subscription,
-		"test-cluster-resource-group",
-		"test-kubelet-identity-client-id",
-		testOptions.NodeResourceGroup,
-		region,
-		testOptions.ProvisionMode,
-	)
 	loadBalancerProvider := loadbalancer.NewProvider(
 		loadBalancersAPI,
 		loadBalancerCache,
@@ -213,7 +197,7 @@ func NewRegionalEnvironment(ctx context.Context, env *coretest.Environment, regi
 	vmInstanceProvider := instance.NewDefaultVMProvider(
 		azClient,
 		instanceTypesProvider,
-		launchTemplateProvider,
+		imageFamilyResolver,
 		loadBalancerProvider,
 		networkSecurityGroupProvider,
 		unavailableOfferingsCache,
@@ -223,6 +207,10 @@ func NewRegionalEnvironment(ctx context.Context, env *coretest.Environment, regi
 		testOptions.ProvisionMode,
 		testOptions.DiskEncryptionSetID,
 		azureEnv,
+		lo.ToPtr("ca-bundle"),
+		testOptions.ClusterEndpoint,
+		"test-tenant",
+		"test-cluster-resource-group",
 	)
 
 	if testOptions.ProvisionMode == consts.ProvisionModeAKSMachineAPI && testOptions.AKSMachinesPoolName != "" {
@@ -286,7 +274,6 @@ func NewRegionalEnvironment(ctx context.Context, env *coretest.Environment, regi
 		KubernetesVersionProvider:    kubernetesVersionProvider,
 		ImageProvider:                imageFamilyProvider,
 		ImageResolver:                imageFamilyResolver,
-		LaunchTemplateProvider:       launchTemplateProvider,
 		LoadBalancerProvider:         loadBalancerProvider,
 		NetworkSecurityGroupProvider: networkSecurityGroupProvider,
